refactor: rename YOCO layer module
