+++ fla/layers/__init__.py
@@ -37,7 +37,7 @@ from .rebased import ReBasedLinearAttention
 from .rodimus import RodimusAttention, SlidingWindowSharedKeyAttention
 from .rwkv6 import RWKV6Attention
 from .rwkv7 import RWKV7Attention
-from .yoco_attn import YOCOCrossAttention, YOCOGatedRetention, YOCOSharedKVBuilder
+from .yoco import YOCOCrossAttention, YOCOGatedRetention, YOCOSharedKVBuilder
 
 __all__ = [
     'ABCAttention',
